requirements.txt
apache-airflow==2.3.0
apache-airflow-providers-ftp==2.1.2   
apache-airflow-providers-http==2.1.2  
apache-airflow-providers-imap==2.2.3  
apache-airflow-providers-sqlite==2.1.3
boto3==1.22.12
botocore==1.25.12
jmespath==0.10.0
jsonpath-ng==1.5.3
jsonschema==4.5.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
nest-asyncio==1.5.5
numpy==1.21.6
openpyxl==3.0.9
opensearch-py==1.1.0
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
pendulum==2.1.2
pg8000==1.26.1
psycopg2==2.9.3
py4j==0.10.9.3
pyarrow==7.0.0
pytz==2022.1
pytzdata==2020.1
requests==2.27.1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
typed-ast==1.5.2
typing-extensions==4.0.1
zipp==3.8.0